move the `io` helpers for reading from file or stdin and writing from file or stdout
